Added Simple Unknown mean Gaussian Ex
Also improved report to handle this. Fixed a bug in reporting objects in
samc.pyx, and took the encoding and decoding and placed it entirely in utils.py

Now have a test_* for quickly running a version of this type of code and exp_*
for actual experiments.

+++ samcnet/samc.pyx
@@ -109,9 +109,9 @@ cdef class SAMCRun:
 
         thetas = self.db['thetas']
         if accessor == None:
-            funcs = self.db['funcs']
+            funcs = self.db['funcs'].astype(np.float)
         else:
-            funcs = np.vectorize(accessor)(self.db['funcs'])
+            funcs = np.vectorize(accessor)(self.db['funcs']).astype(np.float)
 
         part = np.exp(thetas - thetas.max())
         numerator = (part * funcs).cumsum()
@@ -134,9 +134,9 @@ cdef class SAMCRun:
                 "proceeded beyond the burn-in period'
         thetas = self.db['thetas']
         if accessor == None:
-            funcs = self.db['funcs']
+            funcs = self.db['funcs'].astype(np.float)
         else:
-            funcs = np.vectorize(accessor)(self.db['funcs'])
+            funcs = np.vectorize(accessor)(self.db['funcs']).astype(np.float)
 
         if trunc:
             num = len(self.db)/trunc
